Enable the Load Balancer v2 for the Ceilometer(Part One)
In this change set, add the new methods:
list_pools_v2: Support to list the pools with Load Balancer v2 API.
list_members_v2: Support to list the members with Load Balancer v2
API.
list_health_monitors_v2: Support to list the health monitors with
Load Balancer v2 API.

In the Load Balancer v2, the responses of the Pool and Member don't
contian their status. And the status is included in the response of
status_tree[1].

We need to query the status_tree to determine the status of them.

[1]https://wiki.openstack.org/wiki/Neutron/LBaaS/API_2.0

Change-Id: Ibc49282fa39fed864ec9a54d1f7fb38d7034aa9d
Implements: blueprint lbaas-v2-enablement
DocImpact: a new option is introduced