do not configure worker specific items in init
this further corrects some issues with coordination. the basic issue
is described in a bug i fixed a while ago[1]. basically everything
defined in init is sort shared by workers. if it's unique to a worker
it should not be defined in init.

the weird addition to stop() is to maintain bug1418793[2]

[1] I2ad05e2085c0c0f78653c6354d301d18b8dee121
[2] Ied2f086e1f50950b430095ae7ee89036fd4a89d9

Change-Id: I979fdcd350c9a3a0fb6c02e831f1a450133ef76a
Closes-Bug: #1533787