Remove en_GB translations
In preparation for the release, do a cleanup for translations.

Removes all po files that are partially translated. The translation
team has decided to exclude files with less than 66 % of translated
content. There is no content lost, all data is in the translation
server, we just remove it from the repository.

en_GB is only translated to 52 percent.

This change needs to be done manually since the automatic import does
not handle this.

Change-Id: I094fa3f94514b9f42f264badf85e7292016ae9b5